backend/python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
azure-cosmos>=4.7.0
openai>=1.40.0
PyJWT>=2.9.0
